first training done, but model didnt save right :(
