--- requirement.txt
@@ -1,6 +1,0 @@
-numpy
-tensorflow
-kaggle
-matplotlib
-pandas
-python=3.10
