+++ requirements.txt
@@ -1,0 +1,6 @@
+numpy
+tensorflow
+kaggle
+matplotlib
+pandas
+ultralytics
