+++ src/Yolov8/models/yolov8m/results.csv
@@ -1,0 +1,16 @@
+epoch,time,train/box_loss,train/cls_loss,train/dfl_loss,metrics/precision(B),metrics/recall(B),metrics/mAP50(B),metrics/mAP50-95(B),val/box_loss,val/cls_loss,val/dfl_loss,lr/pg0,lr/pg1,lr/pg2
+1,1946.26,1.80809,1.05773,1.42472,0.80402,0.75056,0.82435,0.46826,1.55267,0.89291,1.44748,0.000664694,0.000664694,0.000664694
+2,3847.82,1.70754,0.84412,1.37168,0.83504,0.78113,0.85927,0.47637,1.60042,0.90251,1.46487,0.00124349,0.00124349,0.00124349
+3,5645.73,1.68158,0.8183,1.36234,0.87786,0.79723,0.88709,0.50015,1.54445,0.8152,1.41858,0.00173429,0.00173429,0.00173429
+4,7435.18,1.66368,0.79361,1.35588,0.87467,0.79709,0.87853,0.47767,1.6176,0.80555,1.47215,0.001604,0.001604,0.001604
+5,9225.25,1.62433,0.76362,1.33591,0.89631,0.81932,0.90353,0.46948,1.70874,0.75323,1.52159,0.001472,0.001472,0.001472
+6,10979.3,1.62556,0.75715,1.38787,0.89248,0.80215,0.89188,0.47732,1.64014,0.75499,1.48098,0.00134,0.00134,0.00134
+7,12737.2,1.60233,0.7294,1.38171,0.89814,0.82485,0.90856,0.48714,1.63565,0.72627,1.48983,0.001208,0.001208,0.001208
+8,14492.4,1.58346,0.70953,1.37104,0.90202,0.82145,0.90733,0.47816,1.67394,0.74087,1.50005,0.001076,0.001076,0.001076
+9,16245.3,1.57089,0.69911,1.37737,0.91252,0.84819,0.92341,0.48102,1.68284,0.69426,1.50684,0.000944,0.000944,0.000944
+10,17999,1.55016,0.67868,1.35722,0.91263,0.83472,0.91629,0.47095,1.71251,0.73802,1.52717,0.000812,0.000812,0.000812
+11,19732.6,1.54068,0.66651,1.34958,0.90464,0.84387,0.9165,0.47185,1.68679,0.69883,1.51392,0.00068,0.00068,0.00068
+12,21487.8,1.5223,0.6555,1.34205,0.91715,0.85581,0.92876,0.51028,1.59731,0.66307,1.45375,0.000548,0.000548,0.000548
+13,23241.3,1.51142,0.6422,1.34048,0.91784,0.8541,0.92798,0.49066,1.66543,0.67229,1.50758,0.000416,0.000416,0.000416
+14,24994.7,1.49472,0.63007,1.3333,0.91863,0.84845,0.92539,0.47971,1.67866,0.67721,1.5088,0.000284,0.000284,0.000284
+15,26751.1,1.47851,0.61913,1.32675,0.92343,0.85787,0.93037,0.472,1.72433,0.65692,1.5525,0.000152,0.000152,0.000152
