+++ src/Yolov8/pretrain_yolo.py
@@ -4,9 +4,9 @@ base_model = YOLO('yolov8m.pt')
 
 # Train the model
 base_model.train(
-    data='data/dataset.yaml',
-    epochs=15,
-    batch=16,
+    data='dataset.yaml',
+    epochs=20,
+    batch=8,
     imgsz=640,
     lr0=0.001,
     project='src/Yolov8/models/',
@@ -14,7 +14,7 @@ base_model.train(
 )
 
 base_model.export(format='saved_model',
-                  img_size=640,
+                  imgsz=640,
                   dynamic=False,
                   project='src/Yolov8/models/',
                   name='yolov8m'
